requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pydantic[email]
python-dotenv
jinja2
aiofiles
itsdangerous
starlette